Prepare for large files
The subscriber and publisher changed the way they dealt with files,
before they were using C++ API for IO, but that was slow and there's
some data that reflects that the C API can be as high as 7 times
faster.

So, the pubsub adapted and now are prepared to handle very large
files and stream them directly to the system (without a buffer).